PROJECT
This repository contains code and documentation for a Time Series Analysis project focused on predicting the price of Bitcoin using the open-source forecasting tool, Prophet, developed by Facebook. Time series forecasting is a critical component in understanding cryptocurrency trends, and this project offers a practical application of Prophet for Bitcoin price prediction.